增加商城用户购买商品后,平台根据商户的佣金率抽取费用,佣金存入银盟总代理钱包(现金参数)中、对接向银盛发起代付请求，与回调结果信息处理

+++ silver/silver-shop/silver-shop-api/src/main/java/org/silver/shop/model/system/organization/Merchant.java
@@ -27,7 +27,7 @@ public class Merchant implements Serializable {
 	private String merchantAddress;// 商户地址
 	private String merchantTransport;// 商户商品运输方式
 	private int merchantLevel;// 商户等级
-	private long merchantProfit;// 平台服务费(原分润),按0.0X算
+	private double merchantProfit;// 平台服务费(原分润),按0.0X算
 	private String merchantStatus;// 商户状态：1-启用，2-禁用，3-审核
 	private String createBy;// 创建人
 	private Date createDate;// 创建日期
@@ -152,12 +152,12 @@ public class Merchant implements Serializable {
 	public void setMerchantLevel(int merchantLevel) {
 		this.merchantLevel = merchantLevel;
 	}
-
-	public long getMerchantProfit() {
+	
+	public double getMerchantProfit() {
 		return merchantProfit;
 	}
 
-	public void setMerchantProfit(long merchantProfit) {
+	public void setMerchantProfit(double merchantProfit) {
 		this.merchantProfit = merchantProfit;
 	}
